+++ service/internal/env/env_test.go
@@ -28,7 +28,7 @@ func setup() {
 	for _, envVar := range allEnvVars {
 		os.Unsetenv(envVar)
 	}
-	instance = EnvVars{}
+	instance = Vars{}
 	once = sync.Once{}
 }